Update build process and Dockerfile
Update the build process to use GITHUB_ACTOR/TOKEN, instead of
repository secrets (they're not needed anymore, or shouldn't be).

Also update the Docker image to Focal instead of Xenial, pull the
Erlang source from GitHub instead of erlang.org, and update the Erlang
version to 22.3.4.17 for Riak 3.0.4.

For the entrypoint, use ${!VAR} instead of eval since that seems
a little safer than using eval.

Remove some variables that are no longer used or result in errors on
startup. Some variables are also set in the Dockerfile's env stanza by
default, namely nodename and cookie, since these are most likely to be
overridden.